ENHANCED ENVIRONMENATL PERCEPTIVE AUTONOMOUS AGENT FOR VEHICLE AUTOMATION
The development of an Autonomous Intelligent Agent is one of the biggest
unsolved challenges in Artificial Intelligence. autonomous agent
perceives the environment using sensors and performs actions using
actuators. Observing the environment, the actions performed by the
autonomous agent roots on certain
knowledge. Technologies utilizing sensors enables an autonomous agent to
gain knowledge about the environment. Exploring the environment in the
existing system allows autonomous agents to know about the range of
objects present in the environment. Perceiving the environment from this
perspective lets autonomous agent perform actions based on objects and
behave according to them. This kind of object-centric approach
attenuates the dynamism of autonomous agent. In a strenuous environment,
the autonomous agent should not
be only aware of objects present around but should also be aware of the
ambiance to define its behavior. In this project, Learning technique for
the autonomous agent has been proposed which helps them gain knowledge
about the objects as well as the nature of the environment.